convert data first

+++ switchmate/switchmate.cpp
@@ -154,7 +154,7 @@ void process_data(uint8_t *data, size_t data_len)
     	if (data[0] == SERVICE_DATA)
     	{
         // log the full data message
-        logSwitchmateData(data);
+        logSwitchmateData((int)data);
 
         // last bit of the 7th data piece is the Switchmate status
         setSwitchState(int(data[7]) & 1);
